academy-ui/backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
pydantic>=1.8.2
motor>=2.5.1
python-dotenv>=0.19.1
email-validator>=1.1.3
